feat: Add substringById method to MyString class
The code changes add a new method called `substringById` to the `MyString` class. This method takes an integer `id` and a character array `charArray` as parameters and returns a substring of `charArray` starting from the `id` index. If the `id` is within the bounds of `charArray`, the method returns the substring. Otherwise, it returns an empty string.