feat: Add custom getter and setter for age property in Employee class
The code changes in the `example1.kt` file add a custom getter and setter for the `age` property in the `Employee` class. The getter multiplies the value of the `field` (backing field) by 2, while the setter updates the value of the `field` with the provided `value`. This allows for additional logic and control over the `age` property.